# docs/swagger.yaml
basePath: /
definitions:
...
  models.LoginInput:
    properties:
      email:
        type: string
      password:
        type: string
    type: object
  models.Pelanggan:
    properties:
      alamat:
        type: string
      email:
        type: string
      id:
        type: string
      nama:
        type: string
      no_hp:
        type: string
    required:
    - alamat
    - email
    - nama
    - no_hp
    type: object
...
host: backendmbg-production.up.railway.app
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: |-
    API documentation untuk Backend GROSIR

    **Sistem Login:**
    - Admin: restu129@gmail.com / restu123
    - Kasir: kasir129@gmail.com / 123456
    - Driver: driver129@gmail.com / 123456

    **Authentication:**
    - Semua endpoint (kecuali login) memerlukan Bearer Token
    - Token didapat dari endpoint /auth/login
    - Format: Authorization: Bearer {token}

    **Role Permissions:**
    - Admin: Akses penuh ke semua fitur
    - Kasir: CRUD produk, pelanggan, pembayaran, lihat riwayat
    - Driver: Lihat pembayaran assigned, update status selesai, cetak surat jalan
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Backend MBG API
  version: "1.0"
paths:
  /auth/login:
    post:
      consumes:
      - application/json
      description: Login user dengan email dan password
      parameters:
      - description: Login credentials
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/models.LoginInput'
      produces:
      - application/json
      responses:
        "200":
          description: Login berhasil
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Request tidak valid
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Email atau password salah
          schema:
            additionalProperties: true
            type: object
      summary: Login user
      tags:
      - Authentication
...
  /pelanggan:
    get:
      description: Mengambil semua data pelanggan
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Pelanggan'
            type: array
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Get all customers
      tags:
      - Pelanggan
    post:
      consumes:
      - application/json
      description: Membuat pelanggan baru
      parameters:
      - description: Customer data
        in: body
        name: pelanggan
        required: true
        schema:
          $ref: '#/definitions/models.Pelanggan'
      produces:
      - application/json
      responses:
        "201":
          description: Pelanggan berhasil ditambahkan
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Request tidak valid
          schema:
            additionalProperties: true
            type: object
        "422":
          description: Validasi gagal
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Create customer
      tags:
      - Pelanggan
  /pelanggan/{id}:
    delete:
      description: Hapus pelanggan berdasarkan ID
      parameters:
      - description: Customer ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Pelanggan berhasil dihapus
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Delete customer
      tags:
      - Pelanggan
    get:
      description: Mengambil data pelanggan berdasarkan ID
      parameters:
      - description: Customer ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Pelanggan'
        "404":
          description: Pelanggan tidak ditemukan
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Get customer by ID
      tags:
      - Pelanggan
    put:
      consumes:
      - application/json
      description: Update data pelanggan berdasarkan ID
      parameters:
      - description: Customer ID
        in: path
        name: id
        required: true
        type: string
      - description: Customer data
        in: body
        name: pelanggan
        required: true
        schema:
          $ref: '#/definitions/models.Pelanggan'
      produces:
      - application/json
      responses:
        "200":
          description: Pelanggan berhasil diupdate
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Request tidak valid
          schema:
            additionalProperties: true
            type: object
        "422":
          description: Validasi gagal
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Update customer
      tags:
      - Pelanggan
...
schemes:
- https
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"